requirements.txt
gradio>=4.0.0
google-generativeai==0.8.3
sentence-transformers==3.2.1
numpy==1.26.4
pandas==1.5.3
plotly==5.16.1
pypdf2==3.0.1
openpyxl==3.1.5